+++ src/pyjams/text2plot.py
@@ -50,9 +50,10 @@ History
 
 """
 import time as ptime
-# from .str2tex import str2tex
-from pyjams.str2tex import str2tex
-from jams.romanliterals import int2roman
+from .str2tex import str2tex
+from .romanliterals import int2roman
+# from pyjams.str2tex import str2tex
+# from pyjams.romanliterals import int2roman
 
 
 __all__ = ['abc2plot', 'signature2plot']